Fix recycle bin timestamp check
The timestamp was expected to be in milliseconds but it was in seconds

+++ app/src/main/kotlin/org/fossify/voicerecorder/models/Recording.kt
@@ -4,7 +4,7 @@ data class Recording(
     val id: Int,
     val title: String,
     val path: String,
-    val timestamp: Int,
+    val timestamp: Long,
     val duration: Int,
     val size: Int
 )
